Build system fixes
Previously the build process would separately make both the mapmanager
and examplesimulator modules. Since the examplesimulator has a
dependency on the mapmanager, we need to fold them into the same build
process.
The simplest way to do this is to have a top level Makefile which
identifies both modules and lets the kernel build system handle the
dependencies automatically.

Signed-off-by: Mark Riding <mark.riding@mobica.com>